test/reftest/vendor/blink/fast/dom/DOMException/prototype-object.html

<!DOCTYPE html>
<script src="../../../resources/wpt-test-pre.js"></script>
<script src="../../../resources/js-test-pre.js"></script>
<script>

description("This tests the prototype chain of DOMException objects.")

var e;
try {
  document.appendChild(document);
   // raises a HIERARCHY_REQUEST_ERR
} catch (err) {
  e = err;
}

//shouldBeEqualToString("e.toString()", "HierarchyRequestError: Failed to execute 'appendChild' on 'Node': Nodes of type '#document' may not be inserted inside nodes of type '#document'.");
shouldInclude("e.toString()", "HierarchyRequestError")

shouldBeEqualToString("Object.prototype.toString.call(e)", "[object DOMException]");

// TODO: In this Widget Engine, __proto__ of DOMException should be [object Object].
//shouldBeEqualToString("Object.prototype.toString.call(e.__proto__)", "[object DOMExceptionPrototype]");
shouldBeEqualToString("Object.prototype.toString.call(e.__proto__)", "[object Object]");

// TODO: In this Widget Engine, __proto__.__proto__ of DOMException should be [object Error].
//shouldBeEqualToString("Object.prototype.toString.call(e.__proto__.__proto__)", "[object Object]");
shouldBeEqualToString("Object.prototype.toString.call(e.__proto__.__proto__)", "[object Error]");

//shouldBeEqualToString("e.constructor.toString()", "function DOMException() { [native code] }");
shouldInclude("e.constructor.toString()", "function DOMException()");

shouldBeTrue("e instanceof DOMException");
shouldBeTrue("e instanceof Error");
shouldBe("e.constructor", "window.DOMException");
shouldBe("e.HIERARCHY_REQUEST_ERR", "e.constructor.HIERARCHY_REQUEST_ERR");
shouldBe("e.HIERARCHY_REQUEST_ERR", "3");
shouldBe("e.code", "3");
shouldBeEqualToString("e.name", "HierarchyRequestError");

shouldBeEqualToString("e.message", "Failed to execute 'appendChild' on 'Node': Nodes of type '#document' may not be inserted inside nodes of type '#document'.");

</script>
<script src="../../../resources/js-test-post.js"></script>
<script src="../../../resources/wpt-test-post.js"></script>
